Chess Application
Android Chess application

Allows 1 VS 1

has a functionning but not completed bluetooth connection.

an IA that is based on Alpha Beta Algorithm.

And a 1 vs 1 team play.

N.B. : There is a few bugs in the IA and the bluetooth is not yet completed however the 1VS1 is tested and works fine.